+++ BookShop/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
 
-class ShoppingPageConfig(AppConfig):
+class BookShopConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'shopping_page'
+    name = 'BookShop'
