--- shopping_page/templates/shopping_cart.html
@@ -1,25 +1,0 @@
-{% extends 'base.html' %}
-
-{% block content %}
-    <h1>Shopping Cart Page</h1>
-
-    <div class="cart-items">
-        {% if book_data %}
-            <ul>
-                {% for book_info in book_data %}
-                    <li class="cart-item">
-                        <div class="item-image">
-                            <img src="{{ book_info.thumbnail }}" alt="{{ book_info.title }}">
-                        </div>
-                        <div class="item-details">
-                            <h2>{{ book_info.title }}</h2>
-                            <p>Price: ${{ book_info.price }}</p>
-                        </div>
-                    </li>
-                {% endfor %}
-            </ul>
-        {% else %}
-            <p>Your shopping cart is empty.</p>
-        {% endif %}
-    </div>
-{% endblock %}
