+++ templates/base.html
@@ -34,7 +34,7 @@
                         <ul class="navbar-nav">
                             <li class="nav-link">Book Finds</li>
                             <li class="nav-link">Book Club</li>
-                            <li class="nav-link"><a href="{% url 'shopping_page:shopping_main' %}">Book Shop</a></li>
+                            <li class="nav-link"><a href="{% url 'BookShop:shopping_main' %}">Book Shop</a></li>
                             <li class="nav-link">Book Talk</li>
                         </ul>
                     </div>
